moves noise estimation away from fast_lzero requirement
